fix: scan_count & is_paired handling for faster reconnect (#2)
If the connection to the Roon core drops and has been re-established,
the scan function is not stopped. The reason is that the `is_paired`
flag is only set at the very first Roon core connection, reset at
disconnection, but then never set again.

The `paired_core_id` logic is left as is. My understanding is, that the
API always connects to the same Core again, in case multiple Cores are
available.
If that's not the intention, then the whole logic could be simplified.

Furthermore, the `scan_count` counter is reset after a connection has
been established or closed. This allows for faster reconnections, where
this API is used on devices not having permanent connections to the Roon
Core.
Otherwise only the initial 6 scans are performed every 10 seconds, then
only every 60 seconds. This can lead to poor user experience for use
cases where either the WiFi connection drops, or if the client device is
battery powered and goes into standby.